Backup databases are now created. Backup accounts will keep a running list of every account and its starting balance, incase I remove an account form the main db and want to add it back at a later time.
The prior balances database will be appended every time, as apposed to the main database which will have its information replaced upon update.

The goal is to know use the two databases and present the data in a visually appealing way. Need to collect EOY/EOQ values in the backend but I am otherwise clear to start working on frontend again.